requirements.txt
Jinja==2-2.9.6
MarkupSafe==1.0
Werkzeug==0.12.2
click==6.7
flask-0.12.2
gunicorn==19.7.1
itsdangerous==0.24
